# docs/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
  coinmanage.coinObservedInput:
    description: Input to add/remove coin to/from observed list..
    properties:
      coin:
        description: Coin short name
        example: btc
        type: string
    required:
    - coin
    type: object
  coinmanage.coinPriceOutput:
    description: Output for gotten coin price at timestamp.
    properties:
      coin:
        description: Coin short name
        example: btc
        type: string
      price:
        description: Coin price
        example: "1736500490"
        type: string
      timestamp:
        description: Unix timestamp
        example: 1736500490
        minimum: 0
        type: integer
    required:
    - coin
    - timestamp
    type: object
host: 127.0.0.1:8000
info:
  contact: {}
  description: HTTP API для сбора, хранения и отображения стоимости криптовалют.
  title: Cryptocoin Price API
  version: 1.0.0
paths:
  /currency/add:
    post:
      description: Добавление криптовалюты в список наблюдения.
      operationId: observe-coin
      parameters:
      - description: Название криптовалюты
        in: body
        name: Coin
        required: true
        schema:
          $ref: '#/definitions/coinmanage.coinObservedInput'
      responses:
        "204":
          description: Успешное добавление в список наблюдения
        "400":
          description: Невалидное тело запроса
        "404":
          description: Криптовалюта с таким названием не существует
      summary: Добавление криптовалюты в список наблюдения
      tags:
      - currency
  /currency/price:
    get:
      description: Получение цены криптовалюты.
      operationId: get-coin-price
      parameters:
      - description: Название криптовалюты и время
        in: query
        name: coin
        required: true
        type: string
      - description: Время в UNIX-формате
        in: query
        name: timestamp
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/coinmanage.coinPriceOutput'
        "400":
          description: Невалидное тело запроса
        "404":
          description: Ни одна цена криптовалюты не найдена
      summary: Получение цены криптовалюты
      tags:
      - currency
  /currency/remove:
    delete:
      description: Удаление криптовалюты из списка наблюдения.
      operationId: disable-observe-coin
      parameters:
      - description: Название криптовалюты
        in: body
        name: Coin
        required: true
        schema:
          $ref: '#/definitions/coinmanage.coinObservedInput'
      responses:
        "204":
          description: Успешное добавление в список наблюдения
        "400":
          description: Невалидное тело запроса
        "404":
          description: Криптовалюта с таким названием не была добавлена в список наблюдения
            ранее
      summary: Удаление криптовалюты из списка наблюдения
      tags:
      - currency
produces:
- application/json
schemes:
- http
swagger: "2.0"